remove uthash.h only if it exist

+++ project/step 6-3AC_p2/dependencies.sh
@@ -1,4 +1,4 @@
-rm uthash.h
+[ -e uthash.h ] && rm uthash.h
 sudo apt install flex
 sudo apt install bison
 wget https://raw.githubusercontent.com/troydhanson/uthash/master/src/uthash.h
